refactor: 불필요한 import 문 제거 및 줄바꿈 정렬
